Feature: integrate database (#2)
Integrate Prisma and Planetscale DB into project for persistence

Integrated the data layer into the project using Prisma. I wrote the unit and integration tests for a controller for the Machine model which utilizes the data layer.

The project now has a service layer which is a layer that makes requests to the database. The controller will handle business logic between the service layer and the API requests.